Create BalancedTeam.cpp
C. Balanced Team
time limit per test2 seconds
memory limit per test256 megabytes
inputstandard input
outputstandard output
You are a coach at your local university. There are n students under your supervision, the programming skill of the i-th student is ai.

You have to create a team for a new programming competition. As you know, the more students some team has the more probable its victory is! So you have to create a team with the maximum number of students. But you also know that a team should be balanced. It means that the programming skill of each pair of students in a created team should differ by no more than 5.

Your task is to report the maximum possible number of students in a balanced team.

Input
The first line of the input contains one integer n (1≤n≤2⋅105) — the number of students.

The second line of the input contains n integers a1,a2,…,an (1≤ai≤109), where ai is a programming skill of the i-th student.

Output
Print one integer — the maximum possible number of students in a balanced team.

Examples
inputCopy
6
1 10 17 12 15 2
outputCopy
3
inputCopy
10
1337 1337 1337 1337 1337 1337 1337 1337 1337 1337
outputCopy
10
inputCopy
6
1 1000 10000 10 100 1000000000
outputCopy
1
Note
In the first example you can create a team with skills [12,17,15].

In the second example you can take all students in a team because their programming skills are equal.

In the third example you can create a team consisting of a single student (and you cannot create a team consisting of at least two students).