Create ThanosSort.cpp
A. Thanos Sort
time limit per test1 second
memory limit per test256 megabytes
inputstandard input
outputstandard output
Thanos sort is a supervillain sorting algorithm, which works as follows: if the array is not sorted, snap your fingers* to remove the first or the second half of the items, and repeat the process.

Given an input array, what is the size of the longest sorted array you can obtain from it using Thanos sort?

*Infinity Gauntlet required.

Input
The first line of input contains a single number n (1≤n≤16) — the size of the array. n is guaranteed to be a power of 2.

The second line of input contains n space-separated integers ai (1≤ai≤100) — the elements of the array.

Output
Return the maximal length of a sorted array you can obtain using Thanos sort. The elements of the array have to be sorted in non-decreasing order.

Examples
inputCopy
4
1 2 2 4
outputCopy
4
inputCopy
8
11 12 1 2 13 14 3 4
outputCopy
2
inputCopy
4
7 6 5 4
outputCopy
1
Note
In the first example the array is already sorted, so no finger snaps are required.

In the second example the array actually has a subarray of 4 sorted elements, but you can not remove elements from different sides of the array in one finger snap. Each time you have to remove either the whole first half or the whole second half, so you'll have to snap your fingers twice to get to a 2-element sorted array.

In the third example the array is sorted in decreasing order, so you can only save one element from the ultimate destruction.