Create AlyonaAndANarrowFridge.cpp
B. Alyona and a Narrow Fridge
time limit per test1 second
memory limit per test256 megabytes
inputstandard input
outputstandard output
Alyona has recently bought a miniature fridge that can be represented as a matrix with h rows and 2 columns. Initially there is only one shelf at the bottom of the fridge, but Alyona can install arbitrary number of shelves inside the fridge between any two rows. A shelf is two cells wide, does not occupy any space but separates the inside of the fridge to the lower and upper part.

An example of a fridge with h=7 and two shelves. The shelves are shown in black. The picture corresponds to the first example.
Alyona has n bottles of milk that she wants to put in the fridge. The i-th bottle is ai cells tall and 1 cell wide. She can put a bottle on some shelf if the corresponding space above the shelf is at least as tall as the bottle. She can not put a bottle on top of another bottle (if there is no shelf between them). Two bottles can not share a cell.

Alyona is interested in the largest integer k such that she can put bottles 1, 2, ..., k in the fridge at the same time. Find this largest k.

Input
The first line contains two integers n and h (1≤n≤103, 1≤h≤109) — the number of bottles and the height of the fridge.

The second line contains n integers a1, a2, ..., an (1≤ai≤h) — the heights of the bottles.

Output
Print the single integer k — the maximum integer such that Alyona can put the bottles 1, 2, ..., k in the fridge at the same time. If Alyona can put all bottles in the fridge, print n. It is easy to see that Alyona can always put at least one bottle in the fridge.

Examples
inputCopy
5 7
2 3 5 4 1
outputCopy
3
inputCopy
10 10
9 1 1 1 1 1 1 1 1 1
outputCopy
4
inputCopy
5 10
3 1 4 2 4
outputCopy
5
Note
One of optimal locations in the first example is shown on the picture in the statement.

One of optimal locations in the second example is shown on the picture below.


One of optimal locations in the third example is shown on the picture below.