Create KillEmAll.cpp
B. Kill 'Em All
time limit per test1 second
memory limit per test256 megabytes
inputstandard input
outputstandard output
Ivan plays an old action game called Heretic. He's stuck on one of the final levels of this game, so he needs some help with killing the monsters.

The main part of the level is a large corridor (so large and narrow that it can be represented as an infinite coordinate line). The corridor is divided into two parts; let's assume that the point x=0 is where these parts meet.

The right part of the corridor is filled with n monsters — for each monster, its initial coordinate xi is given (and since all monsters are in the right part, every xi is positive).

The left part of the corridor is filled with crusher traps. If some monster enters the left part of the corridor or the origin (so, its current coordinate becomes less than or equal to 0), it gets instantly killed by a trap.

The main weapon Ivan uses to kill the monsters is the Phoenix Rod. It can launch a missile that explodes upon impact, obliterating every monster caught in the explosion and throwing all other monsters away from the epicenter. Formally, suppose that Ivan launches a missile so that it explodes in the point c. Then every monster is either killed by explosion or pushed away. Let some monster's current coordinate be y, then:

if c=y, then the monster is killed;
if y<c, then the monster is pushed r units to the left, so its current coordinate becomes y−r;
if y>c, then the monster is pushed r units to the right, so its current coordinate becomes y+r.
Ivan is going to kill the monsters as follows: choose some integer point d and launch a missile into that point, then wait until it explodes and all the monsters which are pushed to the left part of the corridor are killed by crusher traps, then, if at least one monster is still alive, choose another integer point (probably the one that was already used) and launch a missile there, and so on.

What is the minimum number of missiles Ivan has to launch in order to kill all of the monsters? You may assume that every time Ivan fires the Phoenix Rod, he chooses the impact point optimally.

You have to answer q independent queries.

Input
The first line contains one integer q (1≤q≤105) — the number of queries.

The first line of each query contains two integers n and r (1≤n,r≤105) — the number of enemies and the distance that the enemies are thrown away from the epicenter of the explosion.

The second line of each query contains n integers xi (1≤xi≤105) — the initial positions of the monsters.

It is guaranteed that sum of all n over all queries does not exceed 105.

Output
For each query print one integer — the minimum number of shots from the Phoenix Rod required to kill all monsters.

Example
inputCopy
2
3 2
1 3 5
4 1
5 2 3 5
outputCopy
2
2
Note
In the first test case, Ivan acts as follows:

choose the point 3, the first monster dies from a crusher trap at the point −1, the second monster dies from the explosion, the third monster is pushed to the point 7;
choose the point 7, the third monster dies from the explosion.
In the second test case, Ivan acts as follows:

choose the point 5, the first and fourth monsters die from the explosion, the second monster is pushed to the point 1, the third monster is pushed to the point 2;
choose the point 2, the first monster dies from a crusher trap at the point 0, the second monster dies from the explosion.